@@ -1,9 +1,9 @@
 name: deep_sleep
-description: A new Flutter project.
-version: 1.0.0+1
+description: Application that helps to meditate and sleep
+version: 1.3.0+5
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   animations: ^2.0.1
@@ -12,21 +12,24 @@ dependencies:
   cached_network_image: ^3.1.0
   carousel_slider: ^4.0.0
   cupertino_icons: ^1.0.3
-  dio: ^4.0.0
-  firebase_auth: ^3.1.1
+  firebase_auth: ^3.3.4
+  firebase_core: ^1.10.6
   fl_chart: ^0.40.0
+  flowder: ^0.2.0
   flutter:
     sdk: flutter
-  flutter_riverpod: ^0.14.0+3
-  flutter_svg: ^0.22.0
+  flutter_facebook_auth: ^3.5.7
+  flutter_svg: ^1.0.0
   google_fonts: ^2.1.0
   google_sign_in: ^5.1.0
   hive_flutter: ^1.1.0
-  # just_audio: ^0.9.13
-  # just_audio_background: ^0.0.1-beta.1
+  jiffy: ^5.0.0
   marquee: ^2.2.0
   page_transition: ^2.0.2
-  # path_provider: ^2.0.5
+  palette_generator: ^0.3.2
+  path_provider: ^2.0.7
+  url_launcher: ^6.0.12
+  webview_flutter: ^3.0.0
 
 flutter_gen:
   integrations:
@@ -34,7 +37,7 @@ flutter_gen:
 
 flutter_icons:
   android: "ic_launcher"
-  ios: true
+  ios: false
   image_path: "res/icon.png"
 
 flutter_native_splash:
@@ -46,9 +49,8 @@ flutter:
     - asset/
 
 dev_dependencies: 
-  alice: ^0.2.4
-  build_runner: ^2.1.4
-  flutter_gen_runner: ^3.1.1
+  build_runner: ^2.1.7
+  flutter_gen_runner: ^4.1.3
   flutter_launcher_icons: ^0.9.2
   flutter_native_splash: ^1.2.1
-  lint: ^1.6.0
+  lint: ^1.8.1
