@@ -3,8 +3,6 @@ export 'dart:io';
 export 'dart:math';
 
 export 'package:animations/animations.dart';
-// export 'package:just_audio/just_audio.dart';
-// export 'package:just_audio_background/just_audio_background.dart';
 export 'package:assets_audio_player/assets_audio_player.dart';
 export 'package:auto_size_text/auto_size_text.dart';
 export 'package:carousel_slider/carousel_slider.dart';
@@ -13,27 +11,29 @@ export 'package:flutter/cupertino.dart' hide RefreshCallback;
 export 'package:flutter/gestures.dart';
 export 'package:flutter/material.dart';
 export 'package:flutter/services.dart';
-// export 'package:flutter_riverpod/flutter_riverpod.dart';
 export 'package:flutter_svg/svg.dart';
 export 'package:google_sign_in/google_sign_in.dart';
 export 'package:hive_flutter/hive_flutter.dart';
+export 'package:jiffy/jiffy.dart';
+export 'package:palette_generator/palette_generator.dart';
+export 'package:url_launcher/url_launcher.dart';
+export 'package:webview_flutter/webview_flutter.dart';
 
 export 'gen/assets.gen.dart';
 export 'screens/dashboard/dashboard.dart';
 export 'screens/experience/experience.dart';
-export 'screens/for_you/for_you.dart';
+export 'screens/for_you.dart';
 export 'screens/learn/learn.dart';
 export 'screens/login.dart';
-export 'screens/notification.dart';
 export 'screens/onboard.dart';
 export 'screens/player.dart';
 export 'screens/rest/rest.dart';
 export 'screens/setting.dart';
-export 'screens/usage.dart';
-export 'utils/config.dart';
+export 'screens/usage/usage.dart';
 export 'utils/constant.dart';
+export 'utils/directory.dart';
 export 'utils/extension.dart';
 export 'utils/hive_helper.dart';
-export 'utils/network.dart';
+export 'utils/stopper.dart';
 export 'utils/theme.dart';
 export 'utils/widget.dart';
